Revert back to twindle cli from twindle-core
