+++ .gitignore
@@ -16,6 +16,9 @@ pids
 *.pid.lock
 yarn.lock
 
+# Confidential email config
+nodemailer.config.js
+
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
