--- twindle-cli/package.json
@@ -1,36 +1,0 @@
-{
-	"name": "twindle",
-	"version": "0.0.1",
-	"description": "This cli fetches a twitter thread and converts it into PDF, ePub, Mobi, etc",
-	"main": "index.js",
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node ."
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/twindle-co/twindle-cli.git"
-	},
-	"author": "",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/twindle-co/twindle-cli/issues"
-	},
-	"homepage": "https://github.com/twindle-co/twindle-cli#readme",
-	"devDependencies": {
-		"@types/puppeteer": "^3.0.4"
-	},
-	"dependencies": {
-		"dotenv": "^8.2.0",
-		"epub-gen": "^0.1.0",
-		"handlebars": "^4.7.6",
-		"kleur": "^4.1.3",
-		"node-calibre": "^2.1.1",
-		"node-fetch": "^2.6.1",
-		"nodemailer": "^6.4.14",
-		"puppeteer": "^5.4.1",
-		"smtp-client": "^0.3.3",
-		"string.prototype.matchall": "^4.0.2",
-		"yargs": "^16.1.0"
-	}
-}
