+++ twindle-cli/twitter/index.js
@@ -10,9 +10,15 @@ async function getTweetsFromURL(url) {
   return response;
 }
 
+/**
+ *
+ * @param {string} tweet_id
+ */
 async function getTweetsFromTweetId(tweet_id) {
   await doTweetLookup(tweet_id);
   return collectTweets();
 }
+
+// getTweetsFromTweetId("1323819296267665409");
 
 module.exports = { getTweetsFromURL, getTweetsFromTweetId };
