+++ twindle-cli/twitter/output/twitter-api-response.json
@@ -1,85 +1,160 @@
-({
+{
   "data": [
     {
-      "id": "1323137099906777088",
-      "created_at": "2020-11-02T05:37:21.000Z",
-      "entities": {
-        "hashtags": [
-          { "start": 34, "end": 57, "tag": "AvatarTheLastAirbender" },
-          { "start": 58, "end": 72, "tag": "LegendOfKorra" }
-        ],
-        "urls": [
-          {
-            "start": 73,
-            "end": 96,
-            "url": "https://t.co/4Q0WBZzmy8",
-            "expanded_url": "https://twitter.com/puruvjdev/status/1323137099906777088/photo/1",
-            "display_url": "pic.twitter.com/4Q0WBZzmy8"
-          }
-        ]
-      },
-      "text": "My current wallpaper. I just love #AvatarTheLastAirbender #LegendOfKorra https://t.co/4Q0WBZzmy8",
-      "conversation_id": "1323137099906777088",
       "context_annotations": [
         {
           "domain": {
-            "id": "3",
-            "name": "TV Shows",
-            "description": "Television shows from around the world"
+            "id": "66",
+            "name": "Interests and Hobbies Category",
+            "description": "A grouping of interests and hobbies entities, like Novelty Food or Destinations"
           },
-          "entity": { "id": "10002926010", "name": "Avatar: The Last Airbender" }
+          "entity": {
+            "id": "824777229892661248",
+            "name": "Generic Food",
+            "description": "Generic Food"
+          }
         },
         {
           "domain": {
-            "id": "3",
-            "name": "TV Shows",
-            "description": "Television shows from around the world"
+            "id": "67",
+            "name": "Interests and Hobbies",
+            "description": "Interests, opinions, and behaviors of individuals, groups, or cultures; like Speciality Cooking or Theme Parks"
           },
-          "entity": { "id": "10024119551", "name": "The Legend of Korra" }
+          "entity": {
+            "id": "1037762759415365633",
+            "name": "Pumpkin",
+            "description": "Pumpkin"
+          }
         }
       ],
-      "source": "Twitter Web App",
-      "attachments": { "media_keys": ["3_1323137073524613120"] },
+      "entities": {
+        "urls": [
+          {
+            "start": 78,
+            "end": 101,
+            "url": "https://t.co/yqUFJ4vnNP",
+            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
+            "display_url": "pic.twitter.com/yqUFJ4vnNP"
+          },
+          {
+            "start": 78,
+            "end": 101,
+            "url": "https://t.co/yqUFJ4vnNP",
+            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
+            "display_url": "pic.twitter.com/yqUFJ4vnNP"
+          },
+          {
+            "start": 78,
+            "end": 101,
+            "url": "https://t.co/yqUFJ4vnNP",
+            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
+            "display_url": "pic.twitter.com/yqUFJ4vnNP"
+          },
+          {
+            "start": 78,
+            "end": 101,
+            "url": "https://t.co/yqUFJ4vnNP",
+            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
+            "display_url": "pic.twitter.com/yqUFJ4vnNP"
+          }
+        ]
+      },
+      "conversation_id": "1323819296267665409",
+      "text": "I watched a squirrel eat a pumpkin for like an hour today and it was magical. https://t.co/yqUFJ4vnNP",
       "lang": "en",
-      "public_metrics": { "retweet_count": 0, "reply_count": 0, "like_count": 6, "quote_count": 1 },
+      "created_at": "2020-11-04T02:48:10.000Z",
+      "source": "Twitter for iPhone",
+      "attachments": {
+        "media_keys": [
+          "3_1323819284926255104",
+          "3_1323819285119225857",
+          "3_1323819285555384325",
+          "3_1323819287425986562"
+        ]
+      },
+      "public_metrics": {
+        "retweet_count": 25,
+        "reply_count": 4,
+        "like_count": 118,
+        "quote_count": 1
+      },
       "possibly_sensitive": false,
-      "author_id": "3436392039"
+      "id": "1323819296267665409",
+      "author_id": "228955392"
     }
   ],
   "includes": {
+    "media": [
+      {
+        "width": 2048,
+        "media_key": "3_1323819284926255104",
+        "type": "photo",
+        "height": 1536,
+        "url": "https://pbs.twimg.com/media/El8moDDXUAAVhWX.jpg"
+      },
+      {
+        "width": 2048,
+        "media_key": "3_1323819285119225857",
+        "type": "photo",
+        "height": 1536,
+        "url": "https://pbs.twimg.com/media/El8moDxX0AEi7q-.jpg"
+      },
+      {
+        "width": 1536,
+        "media_key": "3_1323819285555384325",
+        "type": "photo",
+        "height": 2048,
+        "url": "https://pbs.twimg.com/media/El8moFZXEAUI77Z.jpg"
+      },
+      {
+        "width": 2048,
+        "media_key": "3_1323819287425986562",
+        "type": "photo",
+        "height": 1536,
+        "url": "https://pbs.twimg.com/media/El8moMXWMAIARks.jpg"
+      }
+    ],
     "users": [
       {
+        "location": "Grand Rapids, MI",
+        "name": "Kiri Anne Ryan Stewart 🌶🥒",
+        "username": "ohheykiri",
+        "description": "Often funny, always verbosely sincere non-binary woman & jacket lesbian / She/They / Creative / “Easy to be gay about” / DMs Open / INFP / 4w5 @ohsokirious",
         "verified": false,
-        "created_at": "2015-08-23T07:46:19.000Z",
-        "id": "3436392039",
-        "pinned_tweet_id": "1308682456434868225",
-        "public_metrics": {
-          "followers_count": 292,
-          "following_count": 400,
-          "tweet_count": 3606,
-          "listed_count": 5
-        },
         "protected": false,
-        "description": "18. Web dev for 5 years. Loves orchestral music.\nLove Star Wars, Avengers, Harry Potter.\n\nWhat's the deal with this 👆 cat? 🤔",
-        "location": "Hogwarts, Narnia, Middle Earth",
-        "name": "{/* puruvj.dev */}",
+        "profile_image_url": "https://pbs.twimg.com/profile_images/1320072536714084352/jwlzDKhx_normal.jpg",
+        "url": "https://t.co/umkTzdgMVM",
+        "created_at": "2010-12-21T04:11:13.000Z",
         "entities": {
           "url": {
             "urls": [
               {
                 "start": 0,
                 "end": 23,
-                "url": "https://t.co/gdeY0K6smm",
-                "expanded_url": "https://puruvj.dev",
-                "display_url": "puruvj.dev"
+                "url": "https://t.co/umkTzdgMVM",
+                "expanded_url": "http://bit.ly/2KWHgiC",
+                "display_url": "bit.ly/2KWHgiC"
+              }
+            ]
+          },
+          "description": {
+            "mentions": [
+              {
+                "start": 143,
+                "end": 155,
+                "username": "ohsokirious"
               }
             ]
           }
         },
-        "profile_image_url": "https://pbs.twimg.com/profile_images/1316284643675578369/Ym8iAim8_normal.jpg",
-        "url": "https://t.co/gdeY0K6smm",
-        "username": "puruvjdev"
+        "public_metrics": {
+          "followers_count": 9110,
+          "following_count": 737,
+          "tweet_count": 14883,
+          "listed_count": 43
+        },
+        "id": "228955392"
       }
     ]
   }
-},
+}
