--- twindle-cli/twitter/twitter_responses/response-version2-conversation.json
@@ -1,87 +1,0 @@
-{
-    "data":[
-       {
-          "conversation_id":"1320872258085310466",
-          "lang":"en",
-          "id":"1320872258085310466",
-          "created_at":"2020-10-26T23:37:41.000Z",
-          "possibly_sensitive":false,
-          "text":"Minutes before Amy Coney Barrett is confirmed Supreme Court rules 5-3 to throw out ballots in Wisconsin that are postmarked by Election Day but arrive after. 80,000 ballots were counted in primary that arrived after Election Day but were postmarked by then https://t.co/qEZGQBzpOv",
-          "author_id":"15952856",
-          "entities":{
-             "annotations":[
-                {
-                   "start":15,
-                   "end":31,
-                   "probability":0.9516,
-                   "type":"Person",
-                   "normalized_text":"Amy Coney Barrett"
-                },
-                {
-                   "start":46,
-                   "end":58,
-                   "probability":0.9123,
-                   "type":"Organization",
-                   "normalized_text":"Supreme Court"
-                },
-                {
-                   "start":94,
-                   "end":102,
-                   "probability":0.9795,
-                   "type":"Place",
-                   "normalized_text":"Wisconsin"
-                },
-                {
-                   "start":127,
-                   "end":138,
-                   "probability":0.4046,
-                   "type":"Other",
-                   "normalized_text":"Election Day"
-                },
-                {
-                   "start":216,
-                   "end":227,
-                   "probability":0.5778,
-                   "type":"Other",
-                   "normalized_text":"Election Day"
-                }
-             ],
-             "urls":[
-                {
-                   "start":257,
-                   "end":280,
-                   "url":"https://t.co/qEZGQBzpOv",
-                   "expanded_url":"https://twitter.com/stevenmazie/status/1320870358900158466",
-                   "display_url":"twitter.com/stevenmazie/st…"
-                }
-             ]
-          },
-          "public_metrics":{
-             "retweet_count":10045,
-             "reply_count":937,
-             "like_count":15745,
-             "quote_count":2155
-          },
-          "source":"Twitter Web App",
-          "context_annotations":[
-             {
-                "domain":{
-                   "id":"10",
-                   "name":"Person",
-                   "description":"Named people in the world like Nelson Mandela"
-                },
-                "entity":{
-                   "id":"1314602198932750336",
-                   "name":"Amy Coney Barrett"
-                }
-             }
-          ],
-          "referenced_tweets":[
-             {
-                "type":"quoted",
-                "id":"1320870358900158466"
-             }
-          ]
-       }
-    ]
- }
