+++ twindle-cli/twitter/utils/tweet-info.js
@@ -1,31 +1,33 @@
 const { writeFile } = require("fs").promises;
-let tweets = { 
-    common : {
-      created_at: "",
-      count: "",
-      user: {
-        id: "",
-        username: "",
-        name: "",
-        profile_image_url: "",
-        description: ""
-      }
-    }, 
-    data : [] };
+const tweets = {
+  common: {
+    created_at: "",
+    count: "",
+    user: {
+      id: "",
+      username: "",
+      name: "",
+      profile_image_url: "",
+      description: "",
+    },
+  },
+  data: [],
+};
 
 const addTweet = (tweet) => tweets.data.push(tweet);
 
 const addCommon = (tweet, user) => {
   tweets.common.created_at = tweet.created_at;
   tweets.common.user = { ...user };
-}
+  tweets.common.user.profile_image_url = tweets.common.user.profile_image_url.replace(
+    "_normal.",
+    "."
+  );
+};
 
 async function writeTweets() {
   try {
-    await writeFile(
-      "../output/twitter-api-response.json",
-      JSON.stringify(tweets)
-    );
+    await writeFile("../output/twitter-api-response.json", JSON.stringify(tweets));
   } catch (err) {
     console.error(err);
   }
@@ -36,4 +38,4 @@ function collectTweets() {
   return tweets;
 }
 
-module.exports = { addTweet,  addCommon, writeTweets, collectTweets };
+module.exports = { addTweet, addCommon, writeTweets, collectTweets };
