+++ twindle-cli/twitter/utils/tweet-lookup-request.js
@@ -1,83 +1,11 @@
 const { BEARER_TOKEN } = require("./twitter-api-keys");
-const {
-  ENDPOINT_TO_FETCH_CONVERSATION_ID,
-  TWEET_FIELDS,
-  USER_FIELDS,
-  MEDIA_FIELDS,
-  POLL_FIELDS,
-  PLACE_FIELDS,
-  EXPANSIONS,
-} = require("./twitter-endpoints");
 
 const { UserError } = require("../../helpers/error");
 
-const { getTweetObject, checkIfRequestSuccessful } = require("./tweet-utils");
+const { getTweetObject, checkIfRequestSuccessful, getUrl } = require("./tweet-utils");
 const { processTweetLookup } = require("./tweet-lookup-response");
 const fetch = require("node-fetch");
 const { writeFile } = require("fs").promises;
-
-/**
- * Returns the API endpoint URL from `tweet_id`
- * @param {string} tweet_id
- */
-const getUrl = (tweet_id) =>
-  `${ENDPOINT_TO_FETCH_CONVERSATION_ID}${tweet_id}${TWEET_FIELDS}${EXPANSIONS}${USER_FIELDS}${MEDIA_FIELDS}${PLACE_FIELDS}${POLL_FIELDS}`;
-
-/**
- *
- * @param {string} tweet_id
- */
-async function doTweetLookup(tweet_id) {
-  try {
-    // console.log(getUrl(tweet_id));
-    /** @type {Response} */
-    const response = await fetch(getUrl(tweet_id), getRequestOptions());
-    await processResponse(response);
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-/**
- *
- * @param {fetch.Response} response
- */
-async function processResponse(response) {
-  if (!checkIfRequestSuccessful(response)) {
-    throw new UserError(
-      "request-failed",
-      "Request failed. Check your network and try again"
-    );
-  }
-
-  let responseJSON = await response.json();
-
-  if (isTweetDeleted(responseJSON)) {
-    throw new UserError("tweet-deleted", "Cannot fetch details of this tweet.");
-  }
-
-  let tweet = getTweetObject(responseJSON);
-
-  await writeFile(
-    `${process.cwd()}/twitter/output/twitter-api-response.json`,
-    JSON.stringify(responseJSON),
-    "utf8"
-  );
-
-  if (!isTweetNotOlderThanSevenDays(tweet)) {
-    throw new UserError(
-      "tweet-older-than-7-days",
-      "The tweet must not be older than 7 days."
-    );
-  }
-
-  if (!isProvidedTweetFirstTweetOfTheThread(tweet)) {
-    // This ain't the first tweet of the thread. Find out the first of this thread
-    await doTweetLookup(tweet.conversation_id);
-  } else {
-    await processTweetLookup(responseJSON);
-  }
-}
 
 const getRequestOptions = () => {
   return {
@@ -87,8 +15,7 @@ const getRequestOptions = () => {
   };
 };
 
-const isProvidedTweetFirstTweetOfTheThread = (tweet) =>
-  tweet.id === tweet.conversation_id;
+const isProvidedTweetFirstTweetOfTheThread = (tweet) => tweet.id === tweet.conversation_id;
 
 const isTweetNotOlderThanSevenDays = (tweet) => {
   const currentTime = +new Date();
@@ -100,8 +27,55 @@ const isTweetNotOlderThanSevenDays = (tweet) => {
 
 const isTweetDeleted = (responseJSON) => {
   if (responseJSON.errors === undefined) return false;
-  console.log(responseJSON.errors[0]);
+
   return true;
 };
+
+/**
+ *
+ * @param {string} tweet_id
+ */
+async function doTweetLookup(tweet_id) {
+  try {
+    // console.log(getUrl(tweet_id));
+    /** @type {Response} */
+    const response = await fetch(getUrl(tweet_id), getRequestOptions());
+    // console.log(getUrl(tweet_id))
+    return await processResponse(response);
+  } catch (err) {
+    console.error(err);
+  }
+}
+
+/**
+ * Process the data received from Twitter API
+ * @param {Response} response
+ */
+async function processResponse(response) {
+  if (!checkIfRequestSuccessful(response)) {
+    console.log(await response.json());
+
+    throw new UserError("request-failed", "Request failed. Check your network and try again");
+  }
+
+  let responseJSON = await response.json();
+
+  if (isTweetDeleted(responseJSON)) {
+    throw new UserError("tweet-deleted", "Cannot fetch details of this tweet.");
+  }
+
+  let tweet = getTweetObject(responseJSON);
+
+  if (!isTweetNotOlderThanSevenDays(tweet)) {
+    throw new UserError("tweet-older-than-7-days", "The tweet must not be older than 7 days.");
+  }
+
+  if (!isProvidedTweetFirstTweetOfTheThread(tweet)) {
+    // This ain't the first tweet of the thread. Find out the first of this thread
+    await doTweetLookup(tweet.conversation_id);
+  } else {
+    await processTweetLookup(responseJSON);
+  }
+}
 
 module.exports = { doTweetLookup };
