+++ twindle-cli/twitter/utils/tweet-lookup-response.js
@@ -1,9 +1,6 @@
-const { addTweet, addCommon } = require("./tweet-info");
-const {
-  getTweetObject,
-  getUserObject,
-  createCustomTweet,
-} = require("./tweet-utils");
+const { processMediaFromTweet } = require("./process-tweet-entities");
+const { addTweet, addCommon, collectTweets } = require("./tweet-info");
+const { getTweetObject, getUserObject, createCustomTweet } = require("./tweet-utils");
 const { doTweetsSearch } = require("./tweets-search-request");
 
 /**
@@ -13,9 +10,18 @@ const { doTweetsSearch } = require("./tweets-search-request");
 async function processTweetLookup(responseJSON) {
   let tweet = getTweetObject(responseJSON);
   let user = getUserObject(responseJSON);
+
   addCommon(tweet, user);
+
   addTweet(createCustomTweet(tweet, user));
+
+  await processMediaFromTweet(responseJSON);
+
   await doTweetsSearch(tweet.conversation_id, user.username);
+
+  // console.log(collectTweets().data);
+
+  return collectTweets();
 }
 
 module.exports = {
