--- twindle-core/main.js
@@ -1,3 +1,0 @@
-const { getTweetsFromURL, getTweetsFromTweetId } = require("./index");
-//getTweetsFromURL('https://twitter.com/ReutersScience/status/1321153493223329797');
-getTweetsFromTweetId("1322551796238327809");
