--- twindle-core/output/README.md
@@ -1,1 +1,0 @@
-This folder contains the output json of the tweets fetched from Twitter API
