--- twindle-core/output/twitter-api-response.json
@@ -1,160 +1,0 @@
-{
-  "data": [
-    {
-      "context_annotations": [
-        {
-          "domain": {
-            "id": "66",
-            "name": "Interests and Hobbies Category",
-            "description": "A grouping of interests and hobbies entities, like Novelty Food or Destinations"
-          },
-          "entity": {
-            "id": "824777229892661248",
-            "name": "Generic Food",
-            "description": "Generic Food"
-          }
-        },
-        {
-          "domain": {
-            "id": "67",
-            "name": "Interests and Hobbies",
-            "description": "Interests, opinions, and behaviors of individuals, groups, or cultures; like Speciality Cooking or Theme Parks"
-          },
-          "entity": {
-            "id": "1037762759415365633",
-            "name": "Pumpkin",
-            "description": "Pumpkin"
-          }
-        }
-      ],
-      "entities": {
-        "urls": [
-          {
-            "start": 78,
-            "end": 101,
-            "url": "https://t.co/yqUFJ4vnNP",
-            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
-            "display_url": "pic.twitter.com/yqUFJ4vnNP"
-          },
-          {
-            "start": 78,
-            "end": 101,
-            "url": "https://t.co/yqUFJ4vnNP",
-            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
-            "display_url": "pic.twitter.com/yqUFJ4vnNP"
-          },
-          {
-            "start": 78,
-            "end": 101,
-            "url": "https://t.co/yqUFJ4vnNP",
-            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
-            "display_url": "pic.twitter.com/yqUFJ4vnNP"
-          },
-          {
-            "start": 78,
-            "end": 101,
-            "url": "https://t.co/yqUFJ4vnNP",
-            "expanded_url": "https://twitter.com/ohheykiri/status/1323819296267665409/photo/1",
-            "display_url": "pic.twitter.com/yqUFJ4vnNP"
-          }
-        ]
-      },
-      "conversation_id": "1323819296267665409",
-      "text": "I watched a squirrel eat a pumpkin for like an hour today and it was magical. https://t.co/yqUFJ4vnNP",
-      "lang": "en",
-      "created_at": "2020-11-04T02:48:10.000Z",
-      "source": "Twitter for iPhone",
-      "attachments": {
-        "media_keys": [
-          "3_1323819284926255104",
-          "3_1323819285119225857",
-          "3_1323819285555384325",
-          "3_1323819287425986562"
-        ]
-      },
-      "public_metrics": {
-        "retweet_count": 25,
-        "reply_count": 4,
-        "like_count": 118,
-        "quote_count": 1
-      },
-      "possibly_sensitive": false,
-      "id": "1323819296267665409",
-      "author_id": "228955392"
-    }
-  ],
-  "includes": {
-    "media": [
-      {
-        "width": 2048,
-        "media_key": "3_1323819284926255104",
-        "type": "photo",
-        "height": 1536,
-        "url": "https://pbs.twimg.com/media/El8moDDXUAAVhWX.jpg"
-      },
-      {
-        "width": 2048,
-        "media_key": "3_1323819285119225857",
-        "type": "photo",
-        "height": 1536,
-        "url": "https://pbs.twimg.com/media/El8moDxX0AEi7q-.jpg"
-      },
-      {
-        "width": 1536,
-        "media_key": "3_1323819285555384325",
-        "type": "photo",
-        "height": 2048,
-        "url": "https://pbs.twimg.com/media/El8moFZXEAUI77Z.jpg"
-      },
-      {
-        "width": 2048,
-        "media_key": "3_1323819287425986562",
-        "type": "photo",
-        "height": 1536,
-        "url": "https://pbs.twimg.com/media/El8moMXWMAIARks.jpg"
-      }
-    ],
-    "users": [
-      {
-        "location": "Grand Rapids, MI",
-        "name": "Kiri Anne Ryan Stewart 🌶🥒",
-        "username": "ohheykiri",
-        "description": "Often funny, always verbosely sincere non-binary woman & jacket lesbian / She/They / Creative / “Easy to be gay about” / DMs Open / INFP / 4w5 @ohsokirious",
-        "verified": false,
-        "protected": false,
-        "profile_image_url": "https://pbs.twimg.com/profile_images/1320072536714084352/jwlzDKhx_normal.jpg",
-        "url": "https://t.co/umkTzdgMVM",
-        "created_at": "2010-12-21T04:11:13.000Z",
-        "entities": {
-          "url": {
-            "urls": [
-              {
-                "start": 0,
-                "end": 23,
-                "url": "https://t.co/umkTzdgMVM",
-                "expanded_url": "http://bit.ly/2KWHgiC",
-                "display_url": "bit.ly/2KWHgiC"
-              }
-            ]
-          },
-          "description": {
-            "mentions": [
-              {
-                "start": 143,
-                "end": 155,
-                "username": "ohsokirious"
-              }
-            ]
-          }
-        },
-        "public_metrics": {
-          "followers_count": 9110,
-          "following_count": 737,
-          "tweet_count": 14883,
-          "listed_count": 43
-        },
-        "id": "228955392"
-      }
-    ]
-  }
-}
